==> essay1/R code/00.libs.R <==
# bibliotecas para obtenção dos dados
library(ipeadatar)
library(rbcb)
library(sidrar)
library(quantmod)

# biblioteca para leitura de dados
library(here)
library(readxl)
library(stringr)

# bibliotecas para análise séries temporais
library(urca)
library(mFilter)
library(tseries)
library(forecast)
library(seasonal)
library(tempdisagg)
library(aod)
library(xts)
library(zoo) # for cbind.zoo
library(lubridate)
library(fUnitRoots)

# bibliotecas de formatação das saídas
library(knitr)
library(corrplot)
library(kableExtra)
library(ggplot2)
library(glue)
library(gtools)
library(stringi)
library(pander)
library(gridExtra)
library(grid)
library(R.utils)
library(xtable)
library(huxtable)

# bibliotecas para manipulação dos dados
library(tidyr)
library(dplyr)
library(tsibble)
library(tidyverse)
library(timetk)
library(gdata)

# biblioteca para modelagem VAR Bayesiano e ARDL
library(BVAR)
library(BVARverse)
library(vars)
library(ARDL)
library(dynlm)
library(dynamac)
library(dLagM)
library(fUnitRoots)
library(urca)
library(aod)
library(strucchange)
library(coda)
library(urca)
library(tictoc)
library(parallel)


# bibliotecas para análise de performance dos modelos
library(MLmetrics)
options("getSymbols.warning4.0" = FALSE)
